+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.2] - 2025-02-01
+
+### Fixed
+
+- Practice mode crash when finishing level (thanks @Vulm)
+
 ## [4.0.1] - 2025-01-31
 
 ### Fixed